fixed #336 and implemented #304
